Separate number_input_i32 out in separate file
